tidy up naming convention of gem
gem name is openactive (all lower case) and lib filename is openactive too but module name will be OpenActive. This mismatch is down to OpenActive's branding decision not to use a space in their name.
and by convention don't use upper case in gem names (and because it will mess over case-sensitive file systems).